+++ lib/library_helper.rb
@@ -1,0 +1,2 @@
+# helper to hold connections in between library files
+# Abandoned, only Booker requires a ridiculus umber of connections, everything else is pretty self-sustaining
